handle missing value successfully
